make all backend except payment and subscription

+++ src/app/api/courses/[id]/route.ts
@@ -3,41 +3,63 @@ import connectDB from "@/db/index";
 import { NextResponse, NextRequest } from "next/server";
 import { isValidObjectId } from "mongoose";
 import getDataFromToken from "@/helpers/checkAuth";
+import { ApiError } from "@/utils/ApiError";
+
 connectDB();
 
-export async function DELETE(req: NextRequest, { params }: { params: { id: string } }){
+export async function DELETE(req: NextRequest, { params }: { params: { id: string } }) {
     try {
         const id = params.id;
         const user = await getDataFromToken(req);
-        if(!user){
-            throw new Error("User Session expired or not logged in");
+        if (!user) {
+            throw new ApiError(401, "User Session expired or not logged in");
         }
 
-        if(user.isVerified===false){
-            throw new Error("User is not verified");
+        if (user.isVerified === false) {
+            throw new ApiError(401, "User is not verified");
         }
 
-        await Courses.findByIdAndDelete(id);
-        return NextResponse.json({ message: "Course deleted successfully" }, { status: 200 });
-    } catch (error:any) {
+        if (user.role !== "admin") throw new ApiError(401, "only admin can delete course");
+
+        if (!isValidObjectId(id)) throw new ApiError(400, "Invalid course ID");
+
+        const course = await Courses.findByIdAndDelete(id);
+
+        if (!course) {
+            throw new ApiError(500, "Course not deleted");
+        }
+
+        return NextResponse.
+            json({
+                message: "Course deleted successfully"
+            }, {
+                status: 200
+            });
+
+    } catch (error: unknown) {
         console.error("Error deleting course:", error);
-        return NextResponse.json({ message: error.message || "Error deleting course" }, { status: 500 });
+        return NextResponse.
+            json({
+                message: error instanceof ApiError ? error.message : "Error deleting course"
+            }, {
+                status: error instanceof ApiError ? error.statusCode : 500
+            });
     }
 };
 
 export async function GET(req: NextRequest, { params }: { params: { id: string } }) {
     try {
         const user = await getDataFromToken(req);
-        if(!user){
+        if (!user) {
             throw new Error("User Session expired or not logged in");
         }
-        if(user.isVerified===false){
+        if (user.isVerified === false) {
             throw new Error("User is not verified");
         }
 
-        if (!isValidObjectId(params.id)) return NextResponse.json(
-            { message: "Invalid course ID" },
-            { status: 400 });
+        if (!isValidObjectId(params.id)) {
+            throw new ApiError(404, "Invalid course ID");
+        }
 
         const course = Courses.aggregate(
             [
@@ -65,23 +87,28 @@ export async function GET(req: NextRequest, { params }: { params: { id: string }
                     }
                 },
                 {
-                    $lookup:{
-                        from:"practicesets",
-                        localField:"practiceSet",
-                        foreignField:"_id",
-                        as:"practiceset"
+                    $lookup: {
+                        from: "practicesets",
+                        localField: "practiceSet",
+                        foreignField: "_id",
+                        as: "practiceset"
                     }
                 },
                 {
                     $addFields: {
-                        teacher:{
-                            $first:"$teacher"
+                        teacher: {
+                            $first: "$teacher"
                         }
                     }
                 }
 
             ]
         )
+
+        if(!course) {
+            throw new ApiError(404, "Course not found");
+        }
+
         return NextResponse.
             json(
                 {
@@ -93,22 +120,28 @@ export async function GET(req: NextRequest, { params }: { params: { id: string }
                     status: 200
 
                 });
-        
-    } catch (error: any) {
+
+    } catch (error: unknown) {
         console.error("Error creating course:", error);
-        return NextResponse.json({ message: error.message || "Error creating course" }, { status: 500 });
+        return NextResponse.
+            json({
+                message: error instanceof ApiError ? error.message : "Error creating course"
+            }, {
+                status: error instanceof ApiError ? error.statusCode : 500
+
+            });
     }
 }
 
-export async function PUT(req: NextRequest, { params }: { params: { id: string } }) {
+export async function PATCH(req: NextRequest, { params }: { params: { id: string } }) {
     try {
         const user = await getDataFromToken(req);
         if (!user) {
-            
-            throw new Error("User Session expired or not logged in");
+
+            throw new ApiError(401, "User Session expired or not logged in");
         }
-        if(user.isVerified===false){
-            throw new Error("User is not verified");
+        if (user.isVerified === false) {
+            throw new ApiError(401, "User is not verified");
         }
 
         const id = params.id;
@@ -135,8 +168,14 @@ export async function PUT(req: NextRequest, { params }: { params: { id: string }
                 {
                     status: 200
                 });
-    } catch (error: any) {
+    } catch (error: unknown) {
         console.error("Error updating course:", error);
-        return NextResponse.json({ message: error.message || "Error updating course" }, { status: 500 });
+        return NextResponse.
+        json({
+             message: error instanceof ApiError ? error.message : "Error updating course" }, { 
+            status: error instanceof ApiError ? error.statusCode : 500
+         });
     }
 }
+
+
